moved file from routes to services and rearranged the position of icon on home page to the center of td
