+++ src/assets/services/SportSchedule.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
-import { SportScheduleObject } from "../interfaces/SportSchedule";
-import { SportObject } from "../interfaces/Sport";
+import { SportScheduleObject } from "../../interfaces/SportSchedule";
+import { SportObject } from "../../interfaces/Sport";
 
 
 const sportNameURL = 'http://127.0.0.1:8000/api/schedule/sport';
