+++ src/styles/Home.css
@@ -24,7 +24,7 @@ thead {
 
 td {
     border-bottom: 1px solid #d6d6d6;
-    height: 6vh;
+    height: 7.5vh;
     padding-left: 5%;
     font-size: 20px;
 }
@@ -34,7 +34,7 @@ tr th:nth-child(1) {
 }
 
 th:nth-child(n + 2):nth-child(-n + 20) {
-    width: 100px; /* Adjust the width as needed */
+    width: 10vw; /* Adjust the width as needed */
 }
 
 tr th:nth-child(odd) {
@@ -51,6 +51,10 @@ tr td:nth-child(odd) {
 
 tr td:nth-child(even) {
     background-color: #e8e8fa;
+}
+
+tr td.special-id {
+    text-align: center;
 }
 
 .ceremonies-icon-schedule {
